Rearranged scripts
