--- resources/scripts/mapping/artist-mapping.yaml
@@ -1,23 +1,0 @@
-metadata:
-  name: "Video Game Mapping"
-  source_ext: "csv"
-
-cvs:
-  - cv:
-    type: "vg-ont:VideoGameObservableProperty"  # Define your custom ontology or use existing ones
-    uri: "vg:videoGameObservableProperty"
-    instances: ["Game", "Artists"]
-
-root:
-  type: "vg-ont:VideoGame"
-  uri: "vg:videoGame"
-  varids: ["Game"]  # Use a unique identifier for each video game, like "Name" or a combination of fields
-  connections:
-    - p: "rdf:type"
-      o: "vg-ont:VideoGame"
-    - p: "vg-ont:hasArtist"
-      inv: "vg-ont:illustrated"
-      o:
-        type: "vg-ont:Artist"
-        uri: "vg:artist"
-        varids: ["Artists"]
