--- resources/scripts/mapping/voice-actor-mapping.yaml
@@ -1,40 +1,0 @@
-metadata:
-  name: "Video Game Mapping"
-  source_ext: "csv"
-
-cvs:
-  - cv:
-    type: "vg-ont:VideoGameObservableProperty"  # Define your custom ontology or use existing ones
-    uri: "vg:videoGameObservableProperty"
-    instances: ["Game", "Character", "VoiceActor"]
-
-root:
-  type: "vg-ont:VideoGame"
-  uri: "vg:videoGame"
-  varids: ["Game"]  # Use a unique identifier for each video game, like "Name" or a combination of fields
-  connections:
-    - p: "rdf:type"
-      o: "vg-ont:VideoGame"
-    - p: "vg-ont:hasVoiceActor"
-      inv: "vg-ont:voiceActedIn"
-      o:
-        type: "vg-ont:VoiceActor"
-        uri: "vg:voiceActor"
-        varids: ["VoiceActor"]
-        connections:
-          - p: "vg-ont:voiced"
-            o:
-              type: "vg-ont:Character"
-              uri: "vg:character"
-              varids: ["Character"]
-    - p: "vg-ont:hasCharacter"
-      o:
-        type: "vg-ont:Character"
-        uri: "vg:character"
-        varids: ["Character"]
-        connections:
-          - p: "vg-ont:voicedBy"
-            o:
-              type: "vg-ont:VoiceActor"
-              uri: "vg:voiceActor"
-              varids: ["VoiceActor"]
